Refactor mention_author logic to message.reply
